[refactor] final check

+++ backend/src/services/guild/item.ts
@@ -7,11 +7,11 @@ import { executeTransaction } from '../../utils/executeTransaction';
 
 export const getAll = async (missionId: number, AdventurerId: number, isAccepted: boolean) => {
   const items = await ItemModel.getAll(missionId);
-  if (isAccepted) {
+  if (isAccepted && items?.length) {
     const itemRecords = await itemRecordService.getAll(items, AdventurerId);
     return itemRecords;
   }
-  if (items) return items;
+  if (items?.length) return items;
 };
 
 export const create = async (items: Item[] | TemplateItem[], missionId: number) => {
@@ -24,34 +24,35 @@ export const create = async (items: Item[] | TemplateItem[], missionId: number) 
 };
 
 export const update = async (items: Item[] | TemplateItem[], missionId: number) => {
-  if (!items) await ItemModel.deleteAll(missionId);
+  if (!items?.length) await ItemModel.deleteAll(missionId);
   else {
-    const updateItemIds: { id: number; content: string }[] = [];
+    const updateItems: { id: number; content: string }[] = [];
     const createItemContents: string[] = [];
     const deleteItemIds: number[] = [];
-
     await Promise.all(
       items.map(async ({ id, content }) => {
         if (content) {
-          id ? updateItemIds.push(id, content) : createItemContents.push(content);
+          id ? updateItems.push({ id, content }) : createItemContents.push(content);
         } else {
           deleteItemIds.push(id);
         }
       })
     );
-    await executeTransaction(async () => {
-      const newItemId = await ItemModel.createMany(missionId, createItemContents);
-      if (!newItemId) throw new ApplicationError(400);
 
-      await ItemModel.updateMany(updateItemIds);
-      await ItemModel.deleteManyById(deleteItemIds);
+    await executeTransaction(async () => {
+      if (createItemContents?.length) {
+        const newItemId = await ItemModel.createMany(missionId, createItemContents);
+        if (!newItemId) throw new ApplicationError(400);
+      }
+      if (updateItems?.length) await ItemModel.updateMany(updateItems);
+      if (deleteItemIds?.length) await ItemModel.deleteManyById(deleteItemIds);
     });
   }
 };
 
 export const remove = async (missionId: number) => {
   const items = await ItemModel.getAll(missionId);
-  if (!items) return;
+  if (!items?.length) return;
   await itemRecordService.deleteAllByMission(missionId);
   await ItemModel.deleteAll(missionId);
 };
